backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv

passlib[bcrypt]
bcrypt
python-jose[cryptography]

openai
chromadb